+++ changelog.md
@@ -5,6 +5,7 @@
 - Convert backend to an abstract service api, allowing for easier support for more services
 - Fix background breaking when using essential (or another mod with universalcraft) in mc 1.20.5+
 - Fix crash in case of malformed html in markdown
+- Fix localizations not working on 1.12.2 forge
 - Added small bisect hosting ad to the top of the browse page, this is to help support the project and can be disabled in the new settings page
 
 ----------------------------------------------------------------------------------------------------